Add option for site variable in `mkMessage`'s generated messages
The `renderMessage` function takes in the site as its first variable, but for most
cases does nothing with it except pass it to recursive cases. This commit allows
the user to choose and name this site parameter, meaning it can be used in the
various messages that `mkMessage` generates. This can allow for a message to
include content based on some site data, as long as that data is present from the
site data purely. For example, this can be used to change what the name of some
component is between different instances of the same subsite.