Revert "Add option for site variable in `mkMessage`'s generated messages"
This reverts commit bec6151a3b7ddf3953b82a4f9a349cbe97b09546.

This is a misfeature added to allow messages to depend on site data, but
negates the ability of the messages feature to do I18n. For example,
if you had a subsite and two instances of that subsite, you may want
different names for whatever the parent site is. This feature was intended
to allow you to sub in the name of the parent site. However, the subbed
in word would not be correctly internationalised, and thus this breaks
a fundamental feature of messages.

The correct way to do the above is to create multiple messages sets with
different parent site names, and their own sets of language files.